feat: sequelize integration part 8 - finished express and sequelize
