+++ example/express-graphql-sequelize/package.json
@@ -16,7 +16,7 @@
   "author": "AnthonyLzq <sluzquinosa@uni.pe>",
   "license": "MIT",
   "dependencies": {
-    "@graphql-tools/schema": "^9.0.2",
+    "@graphql-tools/schema": "^9.0.3",
     "@sinclair/typebox": "^0.24.35",
     "ajv": "^8.11.0",
     "ajv-formats": "^2.1.1",
@@ -42,7 +42,7 @@
     "@types/express-pino-logger": "^4.0.3",
     "@types/http-errors": "^1.8.2",
     "@types/jest": "^29.0.0",
-    "@types/node": "^18.7.14",
+    "@types/node": "^18.7.15",
     "@types/swagger-ui-express": "^4.1.3",
     "@typescript-eslint/eslint-plugin": "^5.36.1",
     "@typescript-eslint/parser": "^5.36.1",
