+++ example/express-sequelize/package.json
@@ -37,7 +37,7 @@
     "@types/express-pino-logger": "^4.0.3",
     "@types/http-errors": "^1.8.2",
     "@types/jest": "^29.0.0",
-    "@types/node": "^18.7.14",
+    "@types/node": "^18.7.15",
     "@types/swagger-ui-express": "^4.1.3",
     "@typescript-eslint/eslint-plugin": "^5.36.1",
     "@typescript-eslint/parser": "^5.36.1",
