+++ example/fastify-graphql/package.json
@@ -18,7 +18,7 @@
   "dependencies": {
     "@fastify/cors": "^7",
     "@fastify/swagger": "^6",
-    "@graphql-tools/schema": "^9.0.2",
+    "@graphql-tools/schema": "^9.0.3",
     "@sinclair/typebox": "^0.24.35",
     "ajv": "^8.11.0",
     "ajv-formats": "^2.1.1",
@@ -35,7 +35,7 @@
     "@jest/types": "^29.0.2",
     "@types/http-errors": "^1.8.2",
     "@types/jest": "^29.0.0",
-    "@types/node": "^18.7.14",
+    "@types/node": "^18.7.15",
     "@typescript-eslint/eslint-plugin": "^5.36.1",
     "@typescript-eslint/parser": "^5.36.1",
     "axios": "^0.27.2",
