+++ example/fastify-graphql/src/network/server.ts
@@ -1,8 +1,15 @@
 import fastify, { FastifyInstance } from 'fastify'
-
+import { ApolloServer } from 'apollo-server-fastify'
+import {
+  ApolloServerPluginDrainHttpServer,
+  ApolloServerPluginLandingPageGraphQLPlayground,
+  ApolloServerPluginLandingPageDisabled
+} from 'apollo-server-core'
+import { ApolloServerPlugin } from 'apollo-server-plugin-base'
 import { dbConnection } from 'database'
+
+import { mergedSchema as schema } from 'graphQL'
 import { applyRoutes } from './router'
-import { validatorCompiler } from './utils'
 
 const PORT = process.env.PORT ?? 1996
 const ENVIRONMENTS_WITHOUT_PRETTY_PRINT = ['production', 'ci']
@@ -23,7 +30,6 @@ class Server {
   }
 
   #config() {
-    this.#app.register(require('@fastify/cors'), {})
     this.#app.addHook('preHandler', (req, reply, done) => {
       reply.header('Access-Control-Allow-Methods', 'GET, POST, PATCH, DELETE')
       reply.header('Access-Control-Allow-Origin', '*')
@@ -34,7 +40,6 @@ class Server {
       reply.header('x-powered-by', 'Simba.js')
       done()
     })
-    this.#app.setValidatorCompiler(validatorCompiler)
     applyRoutes(this.#app)
   }
 
@@ -42,11 +47,48 @@ class Server {
     this.#connection = await dbConnection(this.#app.log)
   }
 
+  #fastifyAppClosePlugin(): ApolloServerPlugin {
+    const app = this.#app
+
+    return {
+      async serverWillStart() {
+        return {
+          async drainServer() {
+            await app.close()
+          }
+        }
+      }
+    }
+  }
+
   public async start(): Promise<void> {
+    const server = new ApolloServer({
+      schema,
+      plugins: [
+        this.#fastifyAppClosePlugin(),
+        ApolloServerPluginDrainHttpServer({ httpServer: this.#app.server }),
+        process.env.NODE_ENV === 'production'
+          ? ApolloServerPluginLandingPageDisabled()
+          : ApolloServerPluginLandingPageGraphQLPlayground()
+      ],
+      context: (): Context => ({
+        log: this.#app.log
+      })
+    })
+
     try {
+      await server.start()
+      this.#app.register(
+        server.createHandler({
+          path: '/api'
+        })
+      )
       await this.#dbConnection()
       await this.#connection?.connect()
       await this.#app.listen(PORT)
+      this.#app.log.info(
+        `GraphQL server listening at: http://localhost:${PORT}${server.graphqlPath}`
+      )
     } catch (e) {
       console.error(e)
     }
