+++ example/fastify/package.json
@@ -29,7 +29,7 @@
     "@jest/types": "^29.0.2",
     "@types/http-errors": "^1.8.2",
     "@types/jest": "^29.0.0",
-    "@types/node": "^18.7.14",
+    "@types/node": "^18.7.15",
     "@typescript-eslint/eslint-plugin": "^5.36.1",
     "@typescript-eslint/parser": "^5.36.1",
     "axios": "^0.27.2",
