+++ lib/src/functions/api/express.js
@@ -799,7 +799,8 @@ class Server {
 
 const server = new Server()
 
-export { server as Server }`
+export { server as Server }
+`
         : `import { Server as HttpServer } from 'http'
 import express from 'express'
 import cors from 'cors'
@@ -1203,7 +1204,7 @@ const updateUser = async (
 }
 
 const deleteUser = async (
-  { id }: { id: string },
+  { id }: { id: ${dbIsSQL ? 'number' : 'string'} },
   { log }: Context
 ): Promise<string> => {
   try {
@@ -1309,7 +1310,7 @@ const Mutation = {
   },
   deleteUser: async (
     parent: unknown,
-    { id }: { id: string },
+    { id }: { id: ${dbIsSQL ? 'number' : 'string'} },
     context: Context
   ): Promise<string> => {
     const validate = ajv.compile(idSchema)
@@ -1389,11 +1390,11 @@ const getUsers = async (
 }
 
 const getUser = async (
-  { id }: { id: string },
+  { id }: { id: ${dbIsSQL ? 'number' : 'string'} },
   { log }: Context
 ): Promise<UserDTO> => {
   try {
-    const user = (await get(id as string)) as UserDTO | null
+    const user = (await get(id${dbIsSQL ? '' : ' as string'})) as UserDTO | null
 
     if (!user) throw new ApolloError(EFU.NOT_FOUND, 'NOT_FOUND')
 
@@ -1423,7 +1424,7 @@ import { errorHandling, GE } from '../utils'
 const Query = {
   getUser: async (
     parent: unknown,
-    { id }: { id: string },
+    { id }: { id: ${dbIsSQL ? 'number' : 'string'} },
     context: Context
   ): Promise<UserDTO> => {
     const { log } = context
@@ -1462,7 +1463,7 @@ export { Query }
 
 const User = gql\`
   type User {
-    id: ID!
+    id: ${dbIsSQL ? 'Int' : 'ID'}!
     name: String!
     lastName: String!
     createdAt: String!
@@ -1471,7 +1472,7 @@ const User = gql\`
 
   type Query {
     getUsers: [User!]!
-    getUser(id: ID!): User!
+    getUser(id: ${dbIsSQL ? 'Int' : 'ID'}!): User!
   }
 
   input StoreUserInput {
@@ -1480,7 +1481,7 @@ const User = gql\`
   }
 
   input UpdateUserInput {
-    id: String!
+    id: ${dbIsSQL ? 'Int' : 'String'}!
     lastName: String!
     name: String!
   }
@@ -1489,7 +1490,7 @@ const User = gql\`
     storeUser(user: StoreUserInput!): User!
     deleteAllUsers: String
     updateUser(user: UpdateUserInput!): User!
-    deleteUser(id: ID!): String
+    deleteUser(id: ${dbIsSQL ? 'Int' : 'ID'}!): String
   }
 \`
 
@@ -1977,7 +1978,7 @@ const main = async ({ projectName, graphQL, database }) => {
 
   await types({ projectName, graphQL, dbIsSQL })
   await network({ projectName, graphQL, dbIsSQL })
-  await schemas({ projectName, dbIsSQL })
+  await schemas({ projectName, dbIsSQL, graphQL })
 
   if (!graphQL) await services({ projectName, dbIsSQL })
