+++ lib/src/functions/api/fastify.js
@@ -564,8 +564,9 @@ migration()
  * @param {Object} args
  * @param {String} args.projectName
  * @param {String} args.graphQL
+ * @param {String} args.dbIsSQL
  */
-const network = async ({ projectName, graphQL }) => {
+const network = async ({ projectName, graphQL, dbIsSQL }) => {
   const createFoldersCommand = `mkdir ${projectName}/src/network \
 ${projectName}/src/network/routes ${projectName}/src/network/utils ${
     graphQL
@@ -578,6 +579,7 @@ ${projectName}/src/graphQL/models/utils/messages`
   if (platform() === 'win32')
     await exec(createFoldersCommand.replaceAll('/', '\\'))
   else await exec(createFoldersCommand)
+
   const network = {
     index: {
       content: `export * from './routes'
@@ -1248,7 +1250,7 @@ const updateUser = async (
 }
 
 const deleteUser = async (
-  { id }: { id: string },
+  { id }: { id: ${dbIsSQL ? 'number' : 'string'} },
   { log }: Context
 ): Promise<string> => {
   try {
@@ -1354,7 +1356,7 @@ const Mutation = {
   },
   deleteUser: async (
     parent: unknown,
-    { id }: { id: string },
+    { id }: { id: ${dbIsSQL ? 'number' : 'string'} },
     context: Context
   ): Promise<string> => {
     const validate = ajv.compile(idSchema)
@@ -1434,11 +1436,11 @@ const getUsers = async (
 }
 
 const getUser = async (
-  { id }: { id: string },
+  { id }: { id: ${dbIsSQL ? 'number' : 'string'} },
   { log }: Context
 ): Promise<UserDTO> => {
   try {
-    const user = (await get(id as string)) as UserDTO | null
+    const user = (await get(id${dbIsSQL ? '' : ' as string'})) as UserDTO | null
 
     if (!user) throw new ApolloError(EFU.NOT_FOUND, 'NOT_FOUND')
 
@@ -1468,7 +1470,7 @@ import { errorHandling, GE } from '../utils'
 const Query = {
   getUser: async (
     parent: unknown,
-    { id }: { id: string },
+    { id }: { id: ${dbIsSQL ? 'number' : 'string'} },
     context: Context
   ): Promise<UserDTO> => {
     const { log } = context
@@ -1507,7 +1509,7 @@ export { Query }
 
 const User = gql\`
   type User {
-    id: ID!
+    id: ${dbIsSQL ? 'Int' : 'ID'}!
     name: String!
     lastName: String!
     createdAt: String!
@@ -1516,7 +1518,7 @@ const User = gql\`
 
   type Query {
     getUsers: [User!]!
-    getUser(id: ID!): User!
+    getUser(id: ${dbIsSQL ? 'Int' : 'ID'}!): User!
   }
 
   input StoreUserInput {
@@ -1525,7 +1527,7 @@ const User = gql\`
   }
 
   input UpdateUserInput {
-    id: String!
+    id: ${dbIsSQL ? 'Int' : 'String'}!
     lastName: String!
     name: String!
   }
@@ -1534,7 +1536,7 @@ const User = gql\`
     storeUser(user: StoreUserInput!): User!
     deleteAllUsers: String
     updateUser(user: UpdateUserInput!): User!
-    deleteUser(id: ID!): String
+    deleteUser(id: ${dbIsSQL ? 'Int' : 'ID'}!): String
   }
 \`
 
@@ -2019,7 +2021,7 @@ const main = async ({ projectName, graphQL, database }) => {
   const dbIsSQL = database !== 'mongo'
 
   await types({ projectName, graphQL, dbIsSQL })
-  await network({ projectName, graphQL })
+  await network({ projectName, graphQL, dbIsSQL })
   await schemas({ projectName, graphQL, dbIsSQL })
 
   if (!graphQL) await services({ projectName, dbIsSQL })
