+++ lib/src/functions/api/index.js
@@ -75,7 +75,7 @@ DELETE http://localhost:1996/api/user/60e7e3b93b01c1a7aa74cd6b
           }`
         : `DB_URI = ${
             ENVIRONMENTS_WITH_DB_URI.includes(process.env.NODE_ENV)
-              ? process.env.DB_URI
+              ? process.env.MONGO_URI
               : `mongodb://mongo:mongo@mongo:27017/${projectName}`
           }`,
       file: `${projectName}/.env`
