+++ lib/src/functions/tests.js
@@ -191,7 +191,7 @@ describe('Simba.js tests', () => {
 
       test('Should return 200 as status code', async () => {
         const result = await axios.post<GetOneUserDTO>(\`\${BASE_URL}/api\`, {
-          query: \`query getUser($id: ID!) {
+          query: \`query getUser($id: ${dbIsSQL ? 'Int' : 'ID'}!) {
             user: getUser(id: $id) {
               id
               name
@@ -271,7 +271,7 @@ describe('Simba.js tests', () => {
 
       test('Should return 200 as status code', async () => {
         const result = await axios.post<DeleteUserDTO>(\`\${BASE_URL}/api\`, {
-          query: \`mutation deleteUser($id: ID!) {
+          query: \`mutation deleteUser($id: ${dbIsSQL ? 'Int' : 'ID'}!) {
             result: deleteUser(id: $id)
           }\`,
           variables: {
