+++ package.json
@@ -9,11 +9,13 @@
     "help": "node ./bin -h",
     "lint": "eslint --ext js lib/ --fix",
     "example": "node -r dotenv/config ./bin -N local-example -D 'This is a test using fastify' -a AnthonyLzq -e sluzquinosa@uni.pe -l mit -H --tests --ghat -F",
-    "build": "npm run build:mongo && npm run build:postgresql",
+    "build": "npm run build:express:all && npm run build:fastify:all",
     "build:mongo": "npm run rm:mongo && npm run build:express && npm run build:fastify && npm run build:express:graphql && npm run build:fastify:graphql",
     "build:postgresql": "npm run rm:postgresql && npm run build:express:sequelize && npm run build:fastify:sequelize && npm run build:express:graphql:sequelize && npm run build:fastify:graphql:sequelize",
     "build:rest": "npm run build:express && npm run build:fastify && npm run build:express:sequelize && npm run build:fastify:sequelize",
     "build:graphql": "npm run build:express:graphql && npm run build:fastify:graphql && npm run build:express:graphql:sequelize && npm run build:fastify:graphql:sequelize",
+    "build:express:all": "npm run build:express && npm run build:express:graphql && npm run build:express:sequelize && npm run build:express:graphql:sequelize",
+    "build:fastify:all": "npm run build:fastify && npm run build:fastify:graphql && npm run build:fastify:sequelize && npm run build:fastify:graphql:sequelize",
     "build:express": "npm run rm:express && npm run cd:mv:example && node -r dotenv/config ./bin -N example/express -D 'This is a test using express' -a AnthonyLzq -e sluzquinosa@uni.pe -l mit -H --tests --ghat && npm run rm:git:express",
     "build:fastify": "npm run rm:fastify && npm run cd:mv:example && node -r dotenv/config ./bin -N example/fastify -D 'This is a test using fastify' -a AnthonyLzq -e sluzquinosa@uni.pe -l mit -H -F --tests --ghat && npm run rm:git:fastify",
     "build:express:graphql": "npm run rm:express:graphql && npm run cd:mv:example && node -r dotenv/config ./bin -N example/express-graphql -D 'This is a test using express with GraphQL' -a AnthonyLzq -e sluzquinosa@uni.pe -l mit -H -g && npm run rm:git:express:graphql",
@@ -23,11 +25,14 @@
     "build:express:graphql:sequelize": "npm run rm:express:graphql:sequelize && npm run cd:mv:example && node -r dotenv/config ./bin -N example/express-graphql-sequelize -D 'This is a test using express with graphql and sequelize' -a AnthonyLzq -e sluzquinosa@uni.pe -l mit -H -g --tests --ghat -d postgres && npm run rm:git:express:graphql:sequelize",
     "build:fastify:graphql:sequelize": "npm run rm:fastify:graphql:sequelize && npm run cd:mv:example && node -r dotenv/config ./bin -N example/fastify-graphql-sequelize -D 'This is a test using fastify with graphql and sequelize' -a AnthonyLzq -e sluzquinosa@uni.pe -l mit -H -F -g --tests --ghat -d postgres && npm run rm:git:fastify:graphql:sequelize",
     "build:and:test:only": "npm run build && npm run test:only",
+    "test": "npm run test:without:restore && git restore .",
     "test:local": "npm run test:without:restore && git restore .",
     "test:without:restore": "npm run build && npm run test:express:local && npm run test:fastify:local && npm run test:express:graphql:local && npm run test:fastify:graphql:local && npm run test:express:sequelize:local && test:fastify:sequelize:local",
     "test:ci:mongo": "npm run build:mongo && npm run test:express:ci && npm run test:fastify:ci && npm run test:express:graphql:ci && npm run test:fastify:graphql:ci",
     "test:ci:postgresql": "npm run build:postgresql && npm run test:express:sequelize:ci && npm run test:fastify:sequelize:ci",
-    "test:only": "npm run test:express:local && npm run test:fastify:local && npm run test:express:graphql:local && npm run test:fastify:graphql:local",
+    "test:only": "npm run test:express:only && npm run test:fastify:only",
+    "test:express:only": "npm run test:express:local && npm run test:express:graphql:local && npm run test:express:graphql:sequelize:local && npm run test:express:sequelize:local",
+    "test:fastify:only": "npm run test:fastify:local && npm run test:fastify:graphql:local && npm run test:fastify:graphql:sequelize:local && npm run test:fastify:graphql:sequelize:local",
     "test:express:local": "if [ -d \"example/express/node_modules\" ]; then npm run test:local --prefix example/express; else echo 'Can not run the test for express api because the dependencies are not installed'; fi",
     "test:fastify:local": "if [ -d \"example/fastify/node_modules\" ]; then npm run test:local --prefix example/fastify; else echo 'Can not run the test for fastify api because the dependencies are not installed'; fi",
     "test:express:graphql:local": "if [ -d \"example/express-graphql/node_modules\" ]; then npm run test:local --prefix example/express-graphql; else echo 'Can not run the test for express api because the dependencies are not installed'; fi",
